pyproject.toml
[tool.poetry]
name = "dicot"
version = "0.1.0"
description = "Controls Futaba Command-Type Servo motors."
license = "MIT"
authors = ["lanius <lanius@nirvake.org>"]
readme = "README.md"
repository = "https://github.com/lanius/dicot"

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-mock = "*"
flake8 = "*"
tox = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
